requirements.txt
opencv_python==4.9.0.80
Pillow==10.3.0
ultralytics==8.1.20
